r_user_table,r_user_experiment

+++ src/main/java/backend/service/UserService.java
@@ -8,5 +8,5 @@ import backend.entity.User;
 public interface UserService {
     User login(String email, String password);
 
-    int register(String email, String password);
+    long register(String email, String password);
 }
